backend/requirements.txt
discord.py
requests
fastapi
uvicorn[standard]
python-multipart
pandas
pyarrow
python-dotenv
opencv-python
httpx
pydantic
pydantic-settings
numpy
openai
